+++ packages/sapa/src/renderer/dom/VNodeComponentRender.js
@@ -222,17 +222,19 @@ export async function renderVNodeComponent(componentInstance, $container) {
   // fragment 로 들어오는 children 리스트를 일렬로 다룬다.
   template = flatTemplate(template);
 
+  // TODO: MULTI ROOT 를 허용하지 않는다.
   if (isArray(template) && template.length > 1) {
-    console.log(template);
-    throw new Error(
-      [
-        `Error Component - ${componentInstance.sourceName}`,
-        "Template root is not must an array, however You can use Fragment instead of it",
-        "Fragment Samples: ",
-        " <>{list}</> ",
-        " <Fragment>{list}</Fragment>",
-      ].join("\n")
-    );
+    // console.log(template);
+    // template = [createVNodeFragment({ children: template })];
+    // throw new Error(
+    //   [
+    //     `Error Component - ${componentInstance.sourceName}`,
+    //     "Template root is not must an array, however You can use Fragment instead of it",
+    //     "Fragment Samples: ",
+    //     " <>{list}</> ",
+    //     " <Fragment>{list}</Fragment>",
+    //   ].join("\n")
+    // );
   }
 
   const rootTemplate = template[0];