+++ packages/design-tokens/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @elf-framework/design-tokens
 
+## 0.0.142
+
+### Patch Changes
+
+- remove unused TreeView Component
+  support SearchView Component
+  support useOpenSearch hook
+
 ## 0.0.140
 
 ### Patch Changes